Add basic boilerplate with sandbox
Grunt is used to compile SCSS into CSS. It will watch your SCSS files
for changes, and re-compile as needed. The idea is to follow the Web
Components paradigm, keeping all code (HTML+JS+CSS) related to each
component in one spot (i.e. that component's subfolder).

AngularJS has great styleguides (@johnpapa and @toddmotto) re: best
practices for component-based architectures, insofar as HTML and JS
is concerned. However, there's not as much info out there re: how to
best apply this approach to stylesheets. For this project:

 - Keep each component in its own folder (nesting is fine)
 - Create a [component].scss file in each component's folder
 - Import components via <script> tags in index.html's <body>
 - Namespace each component in rw-[component] class (HTML+SCSS)

An example sandbox component has been added to illustrate these
concepts, and to provide a space in which to mock components.

Here's a breakdown of what actually happens:

 1. grunt-sass-globbing recursively @imports all *.scss files from
    subfolders of /app into map.scss

 2. app.scss @imports map.scss

 3. grunt-contrib-sass compiles app.scss into app.css

 4. index.html <link>'s app.css

Thus, namespacing is needed because the order in which component
styles are @imported can change over time. Namespacing example:

    <div class="rw-component">
        <!-- Component HTML here -->
    </div>

    .rw-component {
        // Component SCSS here
    }

This methodology may change later, but it seems to be a good start.
Promising projects: BEM, SMACSS, webpack-contrib/css-loader