Add angular-ui-notification via CDN
https://github.com/alexcrack/angular-ui-notification

This service is super-useful: it gives us a standard way to show
feedback to the user, which we can call from anywhere inside our
app. We don't have to worry about adding alert areas to each of
our forms, and the consistency makes for better UX.

We can also use it to pipe server messages to the user. Later on,
we should add some code to parse Django REST Framework errors
into a standard format, which we can feed to Notification.