Add initial ApiService
This service is used to standardize server calls. It offers some
developer niceties: it allows the admin to define the API URL via
the app's run() call. We will allow future services to be configured
similarly, so that there's a central place where all this config can
be found.

In the future, we might have run() pull configs from a JSON document
that's listed in .gitignore. That way, server admins can add configs
w/o worrying about those deployment-specific settings making it into
version control. There's a lot of redundancy built-in in regards to
(reasonable) default settings, but that might be for the best.

We will also need to add more in the way of error parsing. Django
REST Framework (DRF) has several standard ways of reporting errors,
especially when it comes to form validation. It'd be nice to parse
these into a standard title-message format that could be passed
directly to Notification. For now, only the simplest error format
has been implemented.

Anyway, make sure to pull in ApiService into new services, and use
its methods instead of making direct $http calls, whenever possible.