Add AuthService and login page
This is the big one. AuthService is somewhat greedy: redirection
might be best handled elsewhere. Like ApiService, AuthService is
configured in app's run(): there, we can define the login route,
publically accessible routes, and default post-login redirect.

If a user tries to access a restricted page, AuthService will
redirect them to the login page, and if their login is successful,
it'll then redirect them back to the restricted page they were
requesting. If a user tries to access the login page while already
authenticated, they'll be redirected to the default redirect target.

AuthService also stores the username and token in LocalStorage,
so that logins are persisted in between page refreshes. Whether
other session states are persisted is yet to be determined.

Lastly, AuthService will add an Authorization header to all $http
requests, including those made by ApiService. The value it sets
matches the format expected by the Django REST Framework for
token-based authentication.

In regards to Roundware specifically, this works out rather nicely.
We can use the default `rest_framework_views.obtain_auth_token` DRF
view for the login endpoint, and still be quite certain that only
staff users will be able to login via the API. django-guardian's
AnonymousUser does not have a password set, so they cannot login.
Likewise, anonymous users generated during the app's normal use
will be unable to login, since their usernames and passwords are
randomized. As long as the API routes continue to be secured
server-side with the proper permissions, all's well!