Add Authenticated state with topbar
Here, we create a new ui-router state called `authenticated`. Now,
`authenticated` is an abstract state, which means that users cannot
navigate directly to it via `/authenticated`, so it's not a route
per se. In the future, we can create new "child" states that specify
`authenticated` as their parent, which will automatically add the
topbar defined in `authenticated.html`. Omitting the `url` param
makes it so that ui-router does not prepend `/autenticated` to child
routes. Practically speaking, `authenticated` will have two children:
`projects` (non-abstract state with list of projects) and `project`,
which itself will be an abstract state that adds a sidebar.

This will need further work in terms of styling and action dropdown.