Add Project service
Many endpoints in Roundware require session_id. In certain cases, it
isn't actually used to any major effect, but it does play a role in
determining which LocalizedString to retrieve, i.e. by checking what
language is associated with that Session. For admin purposes, this
is not needed; furthermore, localization falls back to English if
session_id is omitted (see serializers.py).

Regardless, for our present purposes, we will just set session_id
to 1 for all GET requests. We will need to double-check to ensure
that this doesn't break anything, but so far so good. This is the
`session-for-admin-uploads` defined in Roundware's fixtures.

Later, we'll need to add POST and PATCH handling.