Add Projects selection screen
This screen allows the user to select a project for editing. The
most interesting part of this commit is the `ui-sref` syntax in
`projects.html` - this is the syntax we will be using whenever
we need to link the user to a different app state. Avoid using
"normal" href's - they are harder to maintain. Conversely,
`ui-sref`s will update to match changes to `$stateProvider`.

At this commit, you may get errors, since the project.overview
state hasn't been defined yet. We will add it shortly.

This view may need more style work later. Also, this might be
necessary at some point:

Dynamically set value of ui-sref in AngularJS:
https://stackoverflow.com/questions/24349731