Add Project state with sidebar
This acts similar to the `authenticated` state. Users are not meant
to navigate directly to `/project` or `/project/:id`. Rather, other
states will set their `parent` to `project`, which will allow them
to share its sidebar.

More work will be needed to make the sidebar more mobile-friendly.
More items will be added to the sidebar as development progresses.

I neglected to mention in the previous commit that I'm using a
deeply nested directory organization scheme here. If a state is
always a child of some other state, then its folder should be
placed inside its parent's folder. If this gets tiresome, we can
adopt a flatter directory structure: all that will need to be
updated (as of this commit) is index.html and app.route.js