Save data from server in Project service
We need a way to share data across controllers. There are many ways
to do so. I want to avoid setting up `$watch`-es or event listeners,
I don't want to inject data using the state router, and I don't like
the idea of storing the data in shared parent controllers.

Therefore, we will be using services to share data. Essentially, we
need to "bind" our controller's scope variables to objects that are
managed by singleton services. If the service updates its tracked
objects, those changes will be visible to the relevant controllers,
and AngularJS's internal `$watch`-ers will catch them, updating the
relevant views.

Next, we will abstract this behavior to a dedicated DataService and
ensure that we can differentiate between "pristine" and "dirty" data,
so that edits we make in our forms don't cascade to elsewhere in the
app until form is submitted, and the server confirms the save.

Check out these resources for more info:

https://stackoverflow.com/questions/21919962/share-data-between-angularjs-controllers
https://daveceddia.com/sharing-data-between-controllers-best-practice-use-a-service/
https://johnpapa.net/sharing-data-in-an-angular-controller-or-an-angular-service/