Abstract data sharing into CacheFactory
See previous commit for more info. Here, we abstracted the server
response caching parts from `ProjectService` into `CacheFactory`.

Inject `CacheFactory` into each `ResourceService`. Create a new
instance of `CacheFactory.Cache` like so:

    var cache = new CacheFactory.Cache( 'project_id' );

Then, whenever you call an `ApiService` method, resolve that promise
using `cache.update` and `cache.error`:

    ApiService.get( 'projects' ).then( cache.update, cache.error );

`CacheFactory.Cache.update` will figure out whether it's working
with a single resource, or a list of them. Then, return one of these:

    return cache.list();
    return cache.detail( id, editable );

This would be a good time to point out that the cache keeps two arrays:
one for "dirty" objects (i.e. `editable`), and one for "clean" ones.
In your controllers, you can choose which one you want to retrieve:

 vm.project = ProjectService.detail( $stateParams.id, true );

...will retrieve the editable version, for instance. Rule of thumb:
when you are working on a form, use the editable version. If you just
want to display stuff, omit the second arg to use the clean version.
There's a good chance that you might need both in some controllers.
If it comes to that, I might refactor Cache to return a single object
containing `{ clean: {}, dirty: {} }`, and it'll be up the views to
decide which to render. We'll cross that bridge when we come to it.

Lastly, I'll likely abstract `ApiService` and `CacheFactory` into a
`DataService`, so that the `ResourceServices` don't have to worry
about chaining `then()` calls to `Cache`. It'll be done automatically.

I'm also somewhat worried about the possibility of breaking the
`response` chain. Multiple `then()` calls can be made on the same
promise, but I'm not sure whether they operate on the original
`response`, or on the response returned by the previous `then()`
call in the chain. Worth looking into.