Refactor clean and dirty in CacheFactory
`CacheFactory` does not accept an `editable` argument in its methods.
Instead, it will return an object (or array of objects) that contains
`clean` and `dirty` properties, each of which contains a copy of the
original data. It is now the responsibility of the view or the
controller to decide whether which subset to render.

The only downside to this is that creating lists or forms might get
tedious. I kinda like having to be explicit in the views, and having
both "states" of the data accessible to controllers under a single
object.