Abstract ApiService and CacheFactory into DataFactory
`DataFactory` contains a `Collection` object, a new instance of which
should be instantiated by each `ResourceService`, similar to the way
in which `CacheFactory` was used.

Its only purpose, really, is to keep things DRY: there is no longer
a need for each `ResourceService` to make `ApiService` calls, or to
keep its own `Cache`. Its instance of `DataFactory.Collection` will
handle all of that for it.

I'm not completely satisfied with the semantics here. Is this new
service a service, factory, or provider? I mean this in the broader
OOP sense, not in regards to AngularJS, necessarily. I decided to
make it a factory, so that it matches `CacheFactor`, but that does
not seem quite right. Generally, I'm calling services that require
instantiation of one of their members `Factory`, but this might
need more thought. Some relevant links:

https://stackoverflow.com/questions/1849618/difference-between-a-factory-provider-and-a-service
https://stackoverflow.com/questions/15666048/angularjs-service-vs-provider-vs-factory