Send only changed fields on update
Given two objects, this function will return a new object that only
contains properties that changed. This is a very naive implementation.
We might need to pull in an external library for this, but one that
does exactly what we want here is surprisingly hard to find.

I feel uneasy about sending the whole object on update, so I'd like to
only send those fields whose values actually changed. DRF is good about
accepting partial PATCH requests, and this commit allows us to take
advantage of that behavior.

If this causes more trouble than it's worth, I'll revert it.