Unwrap data lists contained in object
Most list endpoints use an array as the root element. Some, however,
provide a wrapper object as the root. This commit gives us the ability
to specify a property for unwrapping when deciding if `Cache.update`
should affect a single datum or the whole data set.

Each `ResourceService` can specify its own `WRAPPER`.