Add stub ability to find TagCategory by id
We need a way to easily query a collection by parameters and output
some field for a view, without re-querying the server every time.
This is not an elegant solution, but it's a start. Later, we can
figure out a way to add filter functions to the prototype arrays
containing this data, or maybe write up some AngularJS filters
for such tasks. Likewise, it'd be nice to have `clean` and `dirty`
accessors on the collection itself, not just individual resources.