Add directive to correctly parse integers in forms
When an `ngModel` gets changed via a form, it will be converted to a
string, even if it was originally a number. This is annoying, and has
the potential to cause trouble server-side, if the server isn't set
to handle string inputs for numeric fields. DRF handles it well, but
better safe than sorry.

This directive should be added as an attribute to a form element that
is meant to contain some sort of number (e.g. `*_id` fields). It'll
work with text inputs, selects, and likely others.

For the tag editing form, I had trouble getting the select to output
the correct `TagCategory`'s name per id. Options were populated
correctly, but the select wouldn't change to show the current value.
Adding this directive fixes the issue.