Bump AngularJS from 1.6.0-rc.1 to 1.6.4
The `integer` directive was causing the following error:

    Uncaught TypeError: Cannot read property '$$phase' of null

This would happen when the controller containing an `integer` element
was destroyed, e.g. when the user navigated to a different screen.

I'm not sure what caused the issue, but upgrading to the newest
version of AngularJS solved it. Everything seems stable.