Add list().clean and list().dirty
I got tired of having to specify `clean` or `dirty` for each datum
in a collection returned by `ResourceService.list()`. Now, we can
"filter" by `clean` or `dirty` on a collection-wide level:

    ResourceService.list().clean
    ResourceService.list().dirty
    ResourceService.list().both

Using `both` will produce the same effect as before this change.
The only downside is that now you *do* have to specify a sub-list
to use. Btw, I attempted to use filters for a similar effect, but
they created an infinite digest cycle when passed to SmartTable's
`st-safe-src`.