Add Edit Tag screen stub
We are taking a new approach here: `EditTagController` will be
injected with an `id` parameter, which it will use to retrieve
the proper tag. Here, we determine the parameter using a route
resolve. Typically, I avoid resolving promises or retrieving
data in route resolves, since it spreads out logic that (IMO)
intuitively belongs in the controller, but it's useful here
since this approach mirrors that which is we will be using
later for the modal framework.