Add modal-href directive
This gives us the ability to open a modal in the same way as we would
use `ui-sref` to open a state. `modal-href` expects `modal-data`, which
defines the data object to inject into the modal's controller. As per
`ModalService.open`, only data that corresponds to a `params` entry in
the modal's definition will be injected.

There was a scope conflict w/ `st-button`, so its `scope` definition
had to be removed. It was only serving a documentation purpose anyway.