Return both promise and cache in DataService
I want the flexibility of choosing to use a promise, or the static
cache, anywhere. The cost is having to specify an(other) extra sub-
object whenever we need just the static content. I think it's worth
it. In anything, it clarifies the context of `clean` and `dirty`.

Most importantly, `update()` now reveals its promise, which will
enable us to show loading animations, close modals on success,
provide `Notifications`, etc.