Pad out .modal-content
This is a bit of a hack. Basically, Bootstrap expects that modal
contents must be wrapped in `.modal-header`, `.modal-body`, and
`.modal-footer`. These elements have padding; without them, modal
contents appear squashed against the modal's border.

However, I like the idea of modal contents being reusable outside
of modals. Currently, `edit-tag` is equally usable as a state and
as a modal, for instance. IMO, semantically, modal sub-elements add
nothing, yet force us to restructure our HTML purely for display.
I don't like that. So we are just going to ignore them, and add
padding to the `.modal-content` explicitly.