Add dedicated cancel button
As much as I'd like to make a component usable both as a state and a
modal, it's unlikely to happen. I tried using `$injector` to make
`$uibModalInstance` an optional injection, but for some reason, it
wasn't available. We could assume an interface, i.e. just inject
*something* with e.g. a close() function, but that seems like too
much hassle. Unless this dual functionality is necessary, we will
assume that all modal views *will* forever be modal views.

I might create a dedicated directory for modals.

https://stackoverflow.com/questions/18542032/optional-dependencies-in-angularjs