Add create functionality to DataService
Unfortunately, this is a departure from the bipartite format returned
by other functions in the DataService. It's not possible to know that
the save will succeed ahead of time. Therefore, we can't really create
a stub object in the cache to be filled out after the server returns
the cannonical data. We could assume it to be the case, but that feels
like too much magic at this juncture. If it becomes necessary, we can
revisit the issue.

For now, just take note that `create()` returns just the promise, not
an object that contains both the promise and the cached (shared) data.
Therefore, if it is necessary to edit the returned data further, and
re-save it after the initial create call, you'll need to call
`detail()` using the newly returned id.

Here, we also exposed `create()` to TagService.