Add "New Tag" modal
I'm not too happy with having to split `new` and `edit` into two
separate controllers. This complicates things, potentially: given
this architecture, it would be quite difficult to allow the user
to create a new entity, and then edit it within the same modal.

We can revisit this by providing an abstracted `save()` in the
DataService, which delegates the task to `create()` or `update()`
depending on the URL structure. Other changes would likely be
required. Alternatively, we can open an edit modal after the
new modal is successfully resolved. For now, this is good.