Add check-list directive
This will allow us to bind to a list of checkbox items:
https://stackoverflow.com/a/14519881/1943591

We will be using this to correlate a Project's list of language_ids
with the list of possible Languages returned by the LanguageService.

The directive has been enhanced with an optional attribute called
`check-integer`, which is meant for a similar purpose as the
`integer` directive. Use it whenever you're working with ints
to prevent them from being auto-cast to strings.