Process embedded models into their services
This is a mid-way commit. Embedded models such as localized strings
will now be processed into their correct data collections. There's
new configuration options on each service that embeds models, which
serve to facilitate the identification of fields containing embedded
models. There's no mechanism to ensure that embedded resources are
saved to the server before the "main" model is saved.