Add mapped fields option to model config
Roundware's API contains several fields that are returned from the
server with a name that's different from the one it expects when
data is submitted back to the server. Notably, localized string
fields are returned with an `_admin` suffix. This is annoying.
This commit fixes this discrepancy by mapping incoming fields
to outgoing ones. Note that this transformation happens whenever
data is recieved from the server, so the outgoing fields are the
cannonical ones: i.e. views should use outgoing field names!

This commit also adds localized inputs to the tags modal for
sanity testing.