Add ability to ignore incoming fields
This is necessary due to many factors coming together. The short
version is that when Roundware returns a field that's named the
same as the `stored` name for `mapped` field, that field would be
overwritten. Given that the fields that cause this are useless for
our needs, we will ignore them to avoid this situation.

Specifically, they are useless because they return localized strings
for the language associated with the current user account, which we
don't care about while working on admin functionality.