Add volume control input to Asset form
We discussed the possibility of using a 0-1 slider, but it sounds
like it's best to keep it as a float input for compatibility.

We should add `float` attribute directive to prevent these inputs
from being serialized into strings, but before that, we need to
fix the `integer` directive so that it doesn't throw errors if the
model is undefined (i.e. loading), then copy its logic for `float`.