Separate st-button into btn-edit and btn-delete
There was an issue with conflicting scopes, esp. w/ modal directives,
which prevented the icon type from being defined correctly. All btns
would thus share the same icon. This is the first table that features
both types of buttons in one place, so it's the first time this issue
became evident. I'm tired of typing out the HTML for these buttons,
so I'd like to have this directive, but now it's been split into two.