Force reset of cached datum properties on update
As usual, this is a compound issue: `roundware-server` recalculates
the `attenuation_border` on create or update, but because it uses a
secondary database call to do so, it seems that this calculation's
results are never included with the response to `PATCH` or `POST`.
To the user, it seems as if their edits are "lagging behind" by one
each time. This is fixed by adding the `refresh` option to the
Speaker model, clientside.

Secondarily, it seems that my data service stack contained a bug:
if the cache contained items not present in the server response,
they would be persisted. This issue became apparent when several
shapes overlapped, causing the server to generate a "compound"
`attenuation_border`; the shapes in the new border would be added
to the cached datum, but the old shapes would also remain. This
has been fixed by force-deleting non-essential own properties from
the cached datum before merging. This retains object references,
but also requires further testing. One nasty side-effect is that
the localized inputs now "blink" briefly during an `update` call.

tl;dr attenuation will update correctly, app-wide testing required