Give each Speaker its own $watch
This fixes the blinking attenuation border on update, which was really
a symptom of a bigger problem. It looked like an update on any one
Speaker would trigger an update on all of them. Watching the Speakers
collection, and dynamically creating watchers as speakers are removed
and added fixes that problem. There might be a built-in way to do this,
but I haven't found it yet.