Track Speakers in object keyed by id for watchers
Before, we used to setup watchers for `vm.speakers.clean[i]` attrs.
Unfortunately, when a speaker is deleted - specifically, a speaker
that is not in the last index of `vm.speakers.clean` - it causes
the `vm.speakers.clean` array to shift, yet the watchers would
still be pointed at the same indexes, so some of them would be
watching the wrong speakers, i.e. whatever speaker "slid" into
each index. All in all, it was a mess. Using an object as an
associative array keyed by speaker id avoids this mess.

All in all, this controller needs major refactoring for readability.