Add all fields to Project Overview screen
There's a lot of repetition going on here. We should consider ab-
stracting these `form-groups` into directives, using the `btn` and
`local-inputs` directives as examples. I briefly attempted to do so
prior to this commit, but discovered that the inputs would flash
`undefined` briefly before loading.

This could be fixed using `$watches` inside the directive, wherein
we could return early if the bound value is undefined. This would
be the preferred solution, since the alternative would be to resolve
`detail()` and `list()` promises in the router, before changing state.
This is possible, but would require a significant overhaul, and it
wouldn't fix the problem for when the same templates are used for
creating new resources (since those fields would still be empty).
We could in that case explicitly define the fields as properties for
the object in the entity-new controllers, but that seems superfluous.