Add New Project modal
Again, this is unconventional, since it uses the template for the
Project overview screen, but the reduction in code repetition is
worth it. I had to change the update() to save() in the Overview
controller to make them mutually compatible. We will need to add
a Cancel button, conditionally, which may push us to change the
semantics of `showOptional` to something more like `isNew`.