Add ability to delete Projects
There might be a way to refactor modals such that we can avoid calling
`ModalService` in the controllers, and just use modal-handler for the
callback. Currently, this isn't possible, since the confirm modal just
passes `true` in the result, but we might be able to "pass-through" the
values provided to `modal-data` to the handler via the `close()` call.