Hide the Release button
Given that the "not editing any speaker" state doesn't currently
offer any special functionality, this button is somewhat pointless,
serving only to confuse the UX. For an alternative solution, consider
changing `btn-primary` to `btn-default` to make it less prominent.

If we ever add the option to draw a shape on the map to start the
"Add New Speaker" process, this button should be restored.