Fix missing styles on .angular-ui-tree-drag
When a node is dragged, it is appended right before the `</body>` tag.
This means that it gets taken out of the scope of its `.rw-screen`
element. We fix this by applying a scoped class to the `ui-tree-nodes`
elements, and targeting that class in our SCSS.

Unfortunately, `<ol/>` and `<li/>` definitions got rather lengthy, and
I had to break them up over multiple lines. This is a bit unconvention-
al, but should aid in readability. No good way to do so elegantly.