Fix blank localized strings
Alright. Bit of background: the `localInputs` directive works by
interspersing ids and objects w/in the same array, for each field
containing Localized Strings. An id represents a string that has
already been saved to server, while an object represents a string
that has yet to be created serverside. The data service takes care
of looping through these fields' arrays, and deciding whether to
update or create each string. If it creates some strings, it will
replace their objects w/in that array with their new ids, before
saving the "parent" model to server.

This adds a layer on top of that.

First, we can now define `blanked` as a model option. `blanked` is
an array of field names. We presume these fields to be strings. If
a field that is listed in the `blanked` array is empty, then we
consider this instance of that model to have been "blanked"

We've defined the `text` field for `LocalizedString` as `blanked`

What does that mean?

It means that we will refuse to create or update that model, and
instead, raise an error. Technically, we reject the promise, and
thus break the promise chain responsible for the save process.

This creates another problem: the "parent" model won't save unless
all of its `embedded` models have saved.

Therefore, we "catch" this error in the parent model, and handle
it by either (1) deleting the previously-saved embedded model, or
(2) extract this object from the embedded field's array, and add
it back later, after saving the parent model.

The user may not notice the effects of (2) most of the time, since
the modal closes immediately after saving, but it will come into play
if the form fails some other serverside validation. For example, try
creating a Tag: fill out the "Tag" field, and add a localized string,
but leave the "Category" blank. Your `Tag` won't save, but your new
localized string will be restored, even though during the process it
was saved-to and deleted-from the server. Without (2), its form field
would have been reset.

The one downside to this approach is that a `blanked` model does not
delete itself. Rather, it depends on a "parent" model to do so for it.

tl;dr we don't need `blank=True` for the `localized_string` field, and
`LocalizedStrings` that have been blanked will be deleted.