Fix issue with default session_id
There was a bug where only Assets where `session_id=1` were returned.
The cause is multilayered, but worthwhile to explain.

This was b/c we were sending `session_id=1` as a query string param
with every `GET` request. We were doing so b/c the `projects` endpoint
returns a `400 Bad Request` response if the `session_id` is omitted.
It normally requires a `session_id` so that it can find the Language
associated with that session, so that it can populate its message
fields with the correct LocalizedStrings.

Naturally, this is irrelevant for the admin: we want all strings assoc.
w/ each resource, so we pass `admin=1` as a default param with all
requests (see `ApiService.get()`). However, some models (e.g. Project)
still require the `session_id` (erroneously). Early on in development,
we set it as a default param, but this was an overreaction.

Relevant commits:
 - a7730a582ca528f7651c23f5b4c3807939d258fb
 - 2b5498196e37e602b13efe416f8953a5db3fb8e5

Now, we can set it on a per-model basis. This required some refactoring
of how `defaultParams` are set in our DataService. Instead of setting
default params wholesale (i.e. as an object), we can now define them
individually using a ModelService's `setDefaultParam` method. I think
that passing `undefined` or `null` as the second argument when calling
this method can be used to unset a previously-set default param. Note
that we also use this mechanism to filter various resources by the
current project's id.

Why `session_id=1` specifically? It's the `session-for-admin-uploads`
defined in `roundware-server` fixtures.

It's also worthwhile to note that the `session_id` check serverside may
be done individually for both list and detail endpoints, i.e. just b/c
it works for `/projects` does not necessarily mean that it'll work for
`/projects/:id`. I tested all the models as best I could, but some
might have slipped through.