Add column for audio length or other size metric
It's currently blank for non-audio assets, but it might be nice to use
it dynamically, depending on the type of asset. Images could display
filesize or megapixels. Text could display character count. Unsure
about video: could be filesize or playback length, either way.

This is somewhat redundant, since many browsers include a length
indicator as part of the HTML5 audio player, but it's still useful,
since it allows us to sort by audiolength. If we do display size
for non-audio assets down the line, this sorting logic will need
to be customized to look beyond `audio_length_in_seconds`.

Lastly, we display time in seconds using AngularJS's `date` filter,
which allows us to adjust the serialization format:

https://docs.angularjs.org/api/ng/filter/date