Enable saving on Edit Asset screen
This needs to be tested more thoroughly, but things seem stable.
The `file` field needs to be nulled out to prevent the server
from thinking that we are uploading a new file. The intent is
that once an Asset is uploaded, its file should not be changed.
It could be deleted, or "unpublished" by setting `submitted` to
`false`, but its underlying file cannot be changed.