Add Envelope field to Asset screen
Note that while the relationship is M2M, we decided that the Envelope
field should *not* offer the ability to select multiple options. This
also requires that we serialize `envelope_ids` back into an array, if
it has been modified by the `<select/>`.

I find it interesting that the `ng-model` binding on the `<select/>`
manages to resolve integers even if they are wrapped in an array.
This might be attributed to the fact that the `<select/>` could have
the `multiple` attribute, and AngularJS doesn't know about that until
it has to analyze the field when its value changes.

Note that if you submit the form immediately after the page loads,
without changing the `Envelope` field, `envelope_ids` will not have
changed, i.e. it will be an array, not an int. That's why we check
to see if it's an array already, before wrapping it.

Anyway, our `create()` and `update()` serialization logic for Assets
is getting complex enough that we should consider moving it to the
Asset model. See the following commit for ideas on implementation:

25033fa54ab0fcc8b48d7af5e453b0b96f19a0b6