Create new Session for each new Envelope
The underlying issue is that Envelopes can (and should) be filtered by
`project_id`; however, that `project_id` is drawn from the `project_id`
of the Session assoc. w/ each Envelope. We cannot set `session_id` to
the default of `1` when creating Envelopes, since per the fixtures,
Session #1 is assoc. w/ Project #1 specifically. Therefore, if the
current Project is not #1, none of the Envelopes assoc. w/ Session #1
would be visible after the filter.

This workaround is slightly suboptimal, since it creates a new Session
for each Envelope (setting the `project_id` of that Session to match
the current Project), but it requires the least changes all around.
An optimization would be to create a single "admin" Session for each
Project, and use its id as the default for new Envelopes. The would
have to be done server-side in order to affect the Django admin.

On that note, if we add an Envelope filter right now, Envelopes created
with the Django admin will not show, since they default to Session #1.
This change does not fix that issue, but it's a start.