Enable saving on New Asset screen
This enables file upload for new Assets. The key was to set the
`Content-Type` header to `undefined`, which causes AngularJS to
default to the browser's choice, which in this case is multipart
formdata with a boundary parameter.

There's some latent bugs. First, it seems that the create endpoint
serverside does not accept an `envelope_ids` array. I have no idea
how it'll behave with Localized Strings or Tags either. We need to
improve the validation in the `AssetService` to reject the promise,
rather than `throw` an error. Lastly, we should validate the MIME
type of the selected file to ensure it matches `media_type`.