Add checkboxes to modify submitted status
We need the `getAsset` method to retrieve the `cache.dirty` version
of each asset. We output `cache.clean` in the table, same as we do
whenever we want to display data as it came from the server, but
when we want to modify something, we should do so on `cache.dirty`

It's tricky to balance the two. The approach used here is somewhat
inefficient, but it is easier to understand and maintain.