Add st-refresh directive
We will be overwriting Smart Table's `pipe()` function with our own,
where we will handle the filtering, sorting, and pagination logic.

Problem is that `pipe()` is usually triggered on page load, when the
controller is ready. We want to prevent that, and trigger it manually,
after we've retrieved Tags from the server. This prevents our app from
having to make individual Tag calls when rendering each Asset.

In order to trigger the `pipe()` function manually, we added a small
directive that will let us broadcast an event in our screen controller,
which will cascade to the directive attached to the table controller,
and trigger its `pipe()` function.