Add pagination to Assets table
Pagination is achieved by passing a custom `pipe()` function to
Smart Table, where we can do serverside calls to handle pagination,
search, filtering, and sorting. We use the `st-refresh` directive
to trigger `pipe()` after Tags are ready instead of on page load
to prevent request spam. We use a custom `paginate()` method on
the data collection, instead of `list()`, which will return not
only the cache, but the pagination-related metadata as well.

Typically, this kind of stuff is done clientside by Smart Table.
In order to avoid affecting the original data array, it makes a
"safe copy" of the data, which it can sort and slice without
affecting the original data set. Because we overwrite the `pipe()`
function, we need to make that copy manually. I might be mistaken,
but I think we can actually remove it without risking affecting
the cached data. I'll look into it as a potential optimization.