Set default for UI Items according to mode
This probably belongs in the model architecturally, but this is the
least convoluted way to implement it. It's probably bad to hardcode
the mode in general, though Roundware makes a pretty clear boolean
distinction between "Speak" and "Listen".

Basically, this will set `default` to `true` when UI Items are being
added to a UI Group whose `ui_mode` is "listen". This means that when
users are selecting recordings for listening, all tags will be selected
by default, but when users are making a recording, they must select
which tags to apply manually.