Remove envelope select from Edit Asset screen
Available options have been reduced to "Keep current envelope" and
"Create new envelope". This was prompted by a problem wherein the
delay caused by loading envelopes for large projects affected user
experience to an extent great enough to justify removing the option.

Discussions with @hburgund suggested that changing an asset's envelope
is a rare enough need that it won't be missed in normal workflows.