Add tag-multi-select directive
The `isteven/angular-multi-select` doesn't use `ng-model`, so we need
to set up several `$watch`-ers to serialize data back and forth. This
is unfortunate, but it would have been necessary anyway, since the
module requires us to manually splice in items into the selectable
list to enable Tag Category groupings.

It's a fractal of bad design in my opinion - this stuff should have
been handled by allowing the user to specify object properties via
directive attributes - but it is quite flexible. It also forces us
to create a wrapper directive for this specific functionality, which
should be reusable elsewhere, e.g. for filtering on the Assets table.